src: add fletcher64 checksum generation
A suitable hashing algorithm for binning of duplicate/unique files has
to fulfill the following properties:
- low collision probability
- high speed hash generation
- fast hash comparision
- arbitrary file lengths

Fletcher64 has a collision probability of below 0.5 for 2^31 files, 0.01
for 2^28 files.
Checksumming speed is about 1.5GBytes/s per core on current generation
Haswell.
Hash comparision is a simple uint64 compare.
Checksumming is on 32bit words only, this is not an issue if the
datastream is padded (e.g. zero-padded).

Signed-off-by: Stefan Brüns <stefan.bruens@rwth-aachen.de>