src: add TRIE helper class for file comparision
Each trie node represents a file chunk of 4096 bytes. Each node is keyed
with a hash value of the corresponding chunk, although the hash value
generation is deferred until the node has at least one sibling.

On insertion files are compared with the leftmost file in the trie. When
the two files differ, the new file is also compared with siblings (which
share the prefix), and if there is no match a new sibling is inserted.

Signed-off-by: Stefan Brüns <stefan.bruens@rwth-aachen.de>